vocode_sarvam_exotel_starter (3)/vocode_sarvam_exotel/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
openai==1.47.0
anthropic==0.38.0
# Install vocode from PyPI or GitHub according to your target
vocode==0.1.0  # if unavailable, use 'pip install git+https://github.com/vocodedev/vocode-python'
